Update 214.java
Java Easy 214

+++ Java/Easy/214.java
@@ -2,7 +2,7 @@ class Main {
     public static void main(String[] args) {
         int a = 5;
         double b = 10.5;
-        int c = a + b;
+        double c = a + b;
         System.out.println(c);
     }
 }
